backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.1
pydantic-settings>=2.0.3
httpx>=0.24.1
requests>=2.31.0